Updates Libraries (#2)
I tried this using the latest bucklescript libs and it did not work. This updates bsb, reason-react, and Parcel.

I also had issues with the `&` forked process running so I added concurrently to run both commands at once.

One caveat, (I think this existed before though) if you have a breaking build and start up the app, it will show an error on the html and terminal but once it's resolved and compiles again, the html error does not go away and it does not seem to pick up any changes until you kill the watcher and re-launch. 

Then once the app is started and working, if there are any compiler errors, it will not show the html error, but once the app compiles it will show the results again (no killing required). 

I'm not quite sure if the html errors are supposed to actually work with Reason... Parcel does mention Reason on their readme but i'm not sure how much support it has.